rename goroutine-pool to pool
